requirements.txt
streamlit==1.28.0
pdfplumber==0.10.0
PyGithub==2.1.1
ollama==0.1.7
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
